Fixed mount, unmount, and remove
The mount and unmount do not need to handle block device mounts, and so the related code was removed. BTRFS subvolumes are integrated into the BTRFS filesystem, so they don't require explicit mounting or unmounting as separate entities. The mount home(root) path and the volume name can simply be returned if the volume exists or an error can be returned. The count logic basically tracks the number of containers with this volume as an active bind mount. Also the remove volume routine should currently only remove the btrfs subvolume if no snapshots exist. The steps in the volume remove routine to remove the mount directory was also removed.